实现正常登录、注册

+++ src/main/java/com/seclogin/dao/SysUserRepository.java
@@ -3,6 +3,6 @@ package com.seclogin.dao;
 import com.seclogin.entity.SysUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SysUserRepository extends JpaRepository<SysUser, Long> {
+public interface SysUserRepository extends JpaRepository<SysUser, Integer> {
     SysUser findByUsername(String username);
 }
